light-functional: Move pytest_runtest_setup() under functional_tests/
- this is also a preparing commit to disable old 'tc' fixture
- this pytest_runtest_setup() hook belongs only to functional tests
- unittests can not reach this hook from functional_tests/ dir
